requirements.txt
fastapi 
uvicorn 
streamlit 
langchain 
langchain-community 
langchain-ollama 
fpdf2 
faiss-cpu 
tavily-python 
pypdf
langchain_text_splitters